backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
qdrant-client==2.7.0
google-generativeai==0.3.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
aiohttp==3.9.1
requests==2.31.0
python-multipart==0.0.6
cors==1.0.1
